Removed duplicate dispatcher, and renamed folder
