+++ refactor/dispatcher/app/api/api_v0/api.py
@@ -19,7 +19,7 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 
-from app.api.api_v1.endpoints import workflow
+from app.api.api_v0.endpoints import workflow
 from fastapi import APIRouter
 
 api_router = APIRouter()
