+++ refactor/dispatcher/app/schemas/workflow.py
@@ -23,7 +23,7 @@ from datetime import datetime
 from enum import Enum
 from typing import Any, Optional, Sequence
 
-from pydantic import BaseModel, HttpUrl
+from pydantic import BaseModel
 
 
 class DispatchResponse(BaseModel):
