+++ refactor/dispatcher/main.py
@@ -20,7 +20,7 @@
 
 from pathlib import Path
 
-from app.api.api_v1.api import api_router
+from app.api.api_v0.api import api_router
 from app.core.config import settings
 from fastapi import FastAPI
